Issue #9: nrr.region._sim_dimension -> nrr.species._has_3d
KappaNEURON has been working with a version of NEURON from the trunk
of its mercurial repo checked out on 23/9/2014. This was a development
version of NEURON 7.4. There have been a number of changes in the API
meaning that KappaNEURON does not work with the stable NEURON 7.4
release. The previous method of noting if we are in a 3D section was
neuron.rxd.rxd.region._sim_dimension. It is replaced by
neuron.rxd.rxd.species._has_3d.