Issue #9: Update initialisation to NEURON 7.4 rxd API
KappaNEURON has been working with a version of NEURON from the trunk
of its mercurial repo checked out on 23/9/2014. This was a development
version of NEURON 7.4. There have been a number of changes in the API
meaning that KappaNEURON does not work with the stable NEURON 7.4
release. One is that initialisation is now carried out using
initializer._do_init(), which calls _do_init() functions of region,
species and reaction objects. Therefore we need to create a _do_init()
function in the KappaNEURON class (which inherits from
GeneralizedReaction), which calls _kn_init(), which was previously the
neuron.h.FInitializeHandler. _kn_init() no longer needs to call
neuron.rxd.rxd.init().